now wirte map in the MainCanvas working
writing the raw leaflet in Main, instead of using props to pass
position in to child vue component, works perfectly,
but I think I should find a way to seperate the code.

now the MainCanvas is too crowded.

maybe I can not use vue component,
instead I can use ES modules to seperate it
since I think the state of leaflet and threejs
should not be modified by vue,
because it's graphic library,
modifing state in vue maybe have some kind of
overhead since the vue it's designed for ui